src/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
azure-ai-inference==1.0.0b4
azure-core==1.31.0
certifi==2024.8.30
charset-normalizer==3.3.2
confluent-kafka==2.5.3
distro==1.9.0
exceptiongroup==1.2.2
fastavro==1.9.7
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.0
jiter==0.5.0
numpy==2.1.2
openai==1.51.0
pandas==2.2.3
py4j==0.10.9.7
pydantic==2.9.2
pydantic_core==2.23.4
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
